ueba-service/requirements.txt
fastapi
uvicorn
pydantic
motor
httpx
python-dotenv
